requirements.txt
fastapi
uvicorn
google-genai
python-multipart
requests
pillow